requirements.txt
# requirements.txt
pandas
numpy
scikit-learn
xgboost
statsmodels
joblib
google-cloud-storage
google-cloud-aiplatform
kfp
google-cloud-resource-manager
pytest
pytest-mock
